backend/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "To-do list app"
authors = ["Michal Pytlos <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
SQLAlchemy = "^1.4.48"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
psycopg2-binary = "^2.9.6"

[tool.poetry.scripts]
init_postgres_db = "todo.scripts.init_db:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
ipython = "^8.13.2"
httpx = "^0.24.1"

[tool.coverage.run]
omit = ["todo/scripts/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
